refactor: タスク情報を取得するhookを作成
ロジック共通化のため
